Created User (entity) class. UserDto moved to separate package (user.dto)
