+++ src/main/java/com/kodilla/ecommercee/cart/dto/CartDto.java
@@ -1,7 +1,7 @@
 package com.kodilla.ecommercee.cart.dto;
 
 import com.kodilla.ecommercee.product.dto.ProductDto;
-import com.kodilla.ecommercee.user.domain.UserDto;
+import com.kodilla.ecommercee.user.dto.UserDto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
