+++ src/main/java/com/kodilla/ecommercee/user/domain/User.java
@@ -1,6 +1,5 @@
 package com.kodilla.ecommercee.user.domain;
 
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -8,8 +7,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class UserDto {
+public class User {
 
     private Long userId;
     private String userName;
