+++ src/main/java/com/kodilla/ecommercee/user/dto/UserDto.java
@@ -1,0 +1,19 @@
+package com.kodilla.ecommercee.user.dto;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class UserDto {
+
+    private Long userId;
+    private String userName;
+    private String status;
+    private Long userKey;
+
+}
